remove previous code in embedHandlers and rename to embedBuilder
