+++ DiscordBot/commands/tournaments/create.ts
@@ -20,7 +20,7 @@ import { cron } from "../../../Server/cron";
 import { CronJobType } from "../../../Interfaces/cron";
 import { StageType } from "../../../Interfaces/stage";
 import { TournamentRoleType, TournamentChannelType, getTournamentChannelTypeRoles, forumTags } from "../../../Interfaces/tournament";
-import { EmbedBuilder } from "../../functions/embedHandlers";
+import { EmbedBuilder } from "../../functions/embedBuilder";
 
 async function run (m: Message | ChatInputCommandInteraction) {
     if (!m.guild || !(m.member!.permissions as Readonly<PermissionsBitField>).has(PermissionFlagsBits.Administrator))
